requirements.txt
selenium>=4.10.0
pandas>=1.5.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests
python-dotenv
